requirements.txt
fastapi>=0.115.5
uvicorn>=0.32.1
python-multipart>=0.0.7
python-dotenv>=1.0.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.1.0
typing-extensions>=4.5.0
logging>=0.5.1.2 